+++ R/graph.R
@@ -10,7 +10,7 @@ dep_graph <- function(root_dir) {
   config <- read_config(root_dir)
   web <- parse_script(dir(file_path(root_dir, config[["src_dir"]]),
                           pattern = R_FILE_PATTERN, full.names = TRUE),
-                      base_dir = root_dir)
+                      root_dir = root_dir)
   deps <- get_deps(web)
   V <- names(deps)
   E <- lapply(deps, names)